backend/requirements.txt
fastapi
uvicorn[standard]
python-binance
pydantic
pydantic-settings
python-dotenv
psutil
websockets # For WebSocket communication (though python-binance might pull it)